Move checkAuthenticated file to another subdir
